Semicolon bug removed
Noticed a bug for the h1 and h2, The bug was a semicolon after the closing tags. Once I removed the bug it completely changed the spacing adjustments that were made beforehand. I reverted back to setting the margins to 29px auto for the h1 and h2. I also changed the margin for the temperature border back to 30px.