Removed City component because it was overcomplicating my code
I removed the city component because I realized I was overcomplicating my code. I was sending a property for the city from the entry point and then also sending a property to the cities component from the search component. Both properties were displaying the default city. I decided to simplify my code I will remove the Cities component, add a h1 in the Search component and use the usestate instead.